# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API",
        "contact": {
            "email": "arbyusman24@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/tasks": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "5325d35634686a28458a46c33dbf99dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Required. Max 50 characters.",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Cuci AC ruang kelas"
                                    },
                                    "description": {
                                        "description": "Required.",
                                        "type": "string",
                                        "example": "Cuci seluruh AC ruang kelas, dari lantai 1 sampai 10."
                                    },
                                    "assignee_id": {
                                        "description": "Optional",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 201
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Task created successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 422
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Validation Error"
                                                },
                                                "error": {
                                                    "properties": {
                                                        "title": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The title field is required."
                                                            }
                                                        },
                                                        "description": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The description field is required."
                                                            }
                                                        },
                                                        "assignee_id": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The selected assignee_id is invalid."
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/tasks/{id}/status": {
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task status",
                "operationId": "96a8db1d74951f516c62e9587826c3f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id"
                                ],
                                "properties": {
                                    "status_id": {
                                        "description": "1: To Do, 2: Doing, 3: Done, 4: Canceled",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Task status updated successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 422
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Validation Error"
                                                },
                                                "error": {
                                                    "properties": {
                                                        "status_id": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The status_id field is required."
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Internal Server Error"
                                                },
                                                "error": {
                                                    "type": "string",
                                                    "example": "Something went wrong."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}